Added support for a kwallet backend for KDE users
The original keytar_posix file is renamed to keytar_gnome and
provides the gnome-keyring backend where supported. keytar_kde
implements the same logic for a kwallet backend. A new
keytar_posix file now acts as a wrapper which detects which
backend to use and forwards requests to the appropriate
backend implementation.

Various environment variables are used to guess which backend
is the most appropriate.

kwallet is only supported for KDE Frameworks 5.

The main bug in this code is that openWallet will write to stderr
via kDebug() when keytar is run as a command line program without
a window associated with it. I'm nor sure how to disable this
logging (other than redirecting stderr).

Addresses issue #17