Modified the posix backend to conditionally compile kwallet support
If the environment variable 'USE_KWALLET' is set to 1, then
kwallet support will be compiled in. If not set to 1, keytar will only
support gnome-keyring.

Note that gnome-keyring support is always available and USE_KWALLET
does not force the use of the kwallet backend at runtime, it simply
makes it available. The decision of which backend to use is still based
on different environment variable.